Moved some things around to make it all more easy to look at
